Add new config to pin GH actions and automatically merge some updates (#12)
Based on determinations in RFD 434, we've decided to always pin GitHub
Actions digests. However, I also observed in my own projects that
pinning caused a ton of noise to be generated, and that automerging
updates resulted in a much smoother experience.

We're not going to make this the default for now, but let's add a config
to pin GitHub Actions digests and automatically merge updates to some
allowlisted ones. This is strictly better than the current situation
because it provides greater legibility into the exact versions of
actions used.

## Currently allowlisted actions

* `github.com/actions/checkout`: owned by GitHub itself.
* Several actions under `github.com/taiki-e`: owned by Taiki Endo, who is a leader in the Rust community and who has done a great job managing these actions (I've been following his work for several years).

This list isn't committal and we can add to or remove from it over time.